dropping an item now requires confirmation

+++ lib/gemwarrior/inventory.rb
@@ -101,13 +101,19 @@ module Gemwarrior
 
     def remove_item(item_name)
       if contains_item?(item_name)
-        items.delete_at(items.map(&:name).index(item_name) || items.length)
-
-        unless self.weapon.nil?
-          self.weapon = nil if self.weapon.name.eql?(item_name)
-        end
+        puts "Are you sure you want to permanently throw away #{item_name}? (Y/N)"
+        answer = gets.downcase.chomp!
         
-        return "The #{item_name} has been thrown on the ground, but far out of reach, and you're much too lazy to go get it now, so it's as good as gone."
+        if answer.eql?('y')
+          items.delete_at(items.map(&:name).index(item_name) || items.length)
+          unless self.weapon.nil?
+            self.weapon = nil if self.weapon.name.eql?(item_name)
+          end
+        
+          return "The #{item_name} has been thrown on the ground, but far out of reach, and you're much too lazy to go get it now, so it's as good as gone."
+        else
+          return "You decide to keep #{item_name} for now."
+        end
       else
         ERROR_ITEM_REMOVE_INVALID
       end
